Removal of unmaintaned/obsolete code
This is done to reduce project to its core. By that, possible interested
developer won't be overwhelmed by complexity of utilities not used in
the game itself. More detailed reasoning is provided per utility

client-TechViewer:
Based on wx-widgets, it introduces whole new framework to the project,
and is providing something core project should either provide, or avoid
at all.

client-console:
Simple debug client obsoleted in this role by client-ai.

client-msg-wx:
Also based on wx-widgets. Was created to ease pains of writing long
posts in the ingame messager. It's better to fix ingame messager, common
to all players, that to workaround it with not-so-accessible utility.

generators:
Duplicate code to one used ingame.

oslauncher:
Current state of the project means its only audience is possible python
poweruser/developer. Thus oslauncher - something targeted purely to the
common users, is not needed, and it's more than possible, that it never
will. General audience is now used to web-based or mobile online games.