Made blackholes always visible
Even though for new player it produces some excitement, not knowing
which part of galaxy he started in, most of the time it's just
annoying obstacle.

By this commit, black hole in the center (as well as any other, of
the same galaxy) is visible to all players, so they know where the
center is.

Fleets can no longer travel to the black hole.