Increased price of TL progressing technologies
Rushing through tech tree and skipping less viable tech levels is now
very safe way to play the game. By increasing this threshold, I want to
make the risk while rushing a bit higher.

Slightly bigger bumb in case of technology to TL3, because this TL is so
useful, it's worth the wait (rush to TL3 is most obvious of the rushes)