StarSystem dialog iconified
Introducing prototype work of new look. By using icons instead
of text labels, the UI gets a bit lighter for user to consume
as it is a bit easier to skim through data that way.

Also moving description to icon / tooltip combo allows us to
have proper translations to languages with long words.