Updated price of ship hulls.
All adjusted hulls were too cheap, in relation to what they do.
Titanium Small Hull - as the hull is light, CP price of design will be quite a bit
lower than that of Advanced Small Hull. This should make it more even.
Medium and Large hulls - natural progression of price is to double with each size
step. By being cheap, it made larger hulls more efficient. And as this already happens
because of Turrets and Pods, there is no reason to push it even further.
Stealth hull - this hull should have specialist role, not to be a workhorse. This small
tweak of price tries to make it more clear.