Make structures build in more organic way
Adjust prices of all structures and make them be build with only
sliver of HP. That way, player does not have to wait for so long
to build the structure, but the pace of the game is not altered.

Current constants: 20 % of the CP, 10 % of HP when created. Buildings
created by ships, and ship modules producing those are not affected.