Implemented Coop and Brawl scenarios
Coop is scenario where players try to fight AI together. Implemented
in Circle3CP, in which Mutant is the enemy. There is no enforcement of
peace as that does not prevent trolling anyway. Sane players set relations
and are good to go. As it is just an advanced type of training, reward is
just a global announcement.

Brawl is deathmatch variant of galaxy, in first iteration represented by
Circle3BP. Three players pit against each other, no mutant, only renegades.
There is no diplomatic victory - you have to remain sole player in the galaxy.
If you win, galaxy switches to single mode, and you become the owner.