Slight refactoring of BookingMngr as preparation for personal bookings
More information is now saved, database is more flexible with it's own
set of IDs (instead of previous galaxy_type serving as an ID). This was
mandatory change for the personal bookings to happen.

Also, as separate change, removal of booking threshold - now galaxy needs
to be full to trigger. It shouldn't make any problems though.