Implemented STL drives
STL drives are similarly priced to the FTL, but are lighter, and more
powerful in the effect of raising Att and Def of ships. STL drives
provide players with new option to utilize slower moving fleets, that
are, on the other hand, more effective in combat.

Nuclear STL is introduced as the ultimate STL drive (for now).